Fix creation of PseudoSession View
Before the Pseudosession was created by mathing the ith session of an assessment with the ith assessment session. This is not neccesarily correct since e.g. the second force Assessment could be conducted on the 3rd assessment session as was the case in the current study. Therefore, the PseudoSession was adapted by mathing the sessionID instead of the ith session.